requirements.txt
Babel==2.1.1
Django==1.9
django-grappelli==2.7.2
django-phonenumber-field==1.0.0
phonenumbers==7.2.2
Pillow==3.0.0
PyMySQL==0.6.7
pytz==2015.7
requests==2.9.1
wheel==0.24.0
